Update from geonode/master

+++ geonode/layers/utils.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 #########################################################################
 #
-# Copyright (C) 2012 OpenPlans
+# Copyright (C) 2016 OSGeo
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -35,6 +35,7 @@ from osgeo import gdal
 from django.contrib.auth import get_user_model
 from django.template.defaultfilters import slugify
 from django.core.exceptions import ObjectDoesNotExist
+from django.core.files.storage import default_storage as storage
 from django.core.files import File
 from django.conf import settings
 from django.db.models import Q
@@ -51,6 +52,8 @@ from geonode.utils import http_client
 import tarfile
 
 from zipfile import ZipFile, is_zipfile
+
+from datetime import datetime
 
 logger = logging.getLogger('geonode.layers.utils')
 
@@ -354,8 +357,9 @@ def extract_tarfile(upload_file, extension='.shp', tempdir=None):
 
 
 def file_upload(filename, name=None, user=None, title=None, abstract=None,
-                keywords=[], category=None, regions=[],
-                skip=True, overwrite=False, charset='UTF-8'):
+                keywords=[], category=None, regions=[], date=None,
+                skip=True, overwrite=False, charset='UTF-8',
+                metadata_uploaded_preserve=False):
     """Saves a layer in GeoNode asking as little information as possible.
        Only filename is required, user and title are optional.
     """
@@ -423,10 +427,17 @@ def file_upload(filename, name=None, user=None, title=None, abstract=None,
 
     # set metadata
     if 'xml' in files:
-        xml_file = open(files['xml'])
+        with open(files['xml']) as f:
+            xml_file = f.read()
         defaults['metadata_uploaded'] = True
+        defaults['metadata_uploaded_preserve'] = metadata_uploaded_preserve
+
         # get model properties from XML
-        vals, regions, keywords = set_metadata(xml_file.read())
+        identifier, vals, regions, keywords = set_metadata(xml_file)
+
+        if defaults['metadata_uploaded_preserve']:
+            defaults['metadata_xml'] = xml_file
+            defaults['uuid'] = identifier
 
         for key, value in vals.items():
             if key == 'spatial_representation_type':
@@ -475,7 +486,8 @@ def file_upload(filename, name=None, user=None, title=None, abstract=None,
     # process the layer again after that by
     # doing a layer.save()
     if not created and overwrite:
-        layer.upload_session.layerfile_set.all().delete()
+        if layer.upload_session:
+            layer.upload_session.layerfile_set.all().delete()
         layer.upload_session = upload_session
         # Pass the parameter overwrite to tell whether the
         # geoserver_post_save_signal should upload the new file or not
@@ -494,13 +506,18 @@ def file_upload(filename, name=None, user=None, title=None, abstract=None,
         if len(regions_resolved) > 0:
             layer.regions.add(*regions_resolved)
 
+    if date is not None:
+        layer.date = datetime.strptime(date, '%Y-%m-%d %H:%M:%S')
+        layer.save()
+
     return layer
 
 
 def upload(incoming, user=None, overwrite=False,
            keywords=(), category=None, regions=(),
            skip=True, ignore_errors=True,
-           verbosity=1, console=None, title=None, private=False):
+           verbosity=1, console=None, title=None, date=None,
+           private=False, metadata_uploaded_preserve=False):
     """Upload a directory of spatial data files to GeoNode
 
        This function also verifies that each layer is in GeoServer.
@@ -585,7 +602,9 @@ def upload(incoming, user=None, overwrite=False,
                                     keywords=keywords,
                                     category=category,
                                     regions=regions,
-                                    title=title
+                                    title=title,
+                                    date=date,
+                                    metadata_uploaded_preserve=metadata_uploaded_preserve
                                     )
                 if not existed:
                     status = 'created'
@@ -598,6 +617,16 @@ def upload(incoming, user=None, overwrite=False,
                                                            "delete_resourcebase", "change_resourcebase_permissions",
                                                            "publish_resourcebase"]}, "groups": {}}
                     layer.set_permissions(perm_spec)
+
+                if getattr(settings, 'SLACK_ENABLED', False):
+                    try:
+                        from geonode.contrib.slack.utils import build_slack_message_layer, send_slack_messages
+                        send_slack_messages(build_slack_message_layer(
+                            ("layer_new" if status == "created" else "layer_edit"),
+                            layer))
+                    except:
+                        print "Could not send slack message."
+
             except Exception as e:
                 if ignore_errors:
                     status = 'failed'
@@ -626,51 +655,59 @@ def upload(incoming, user=None, overwrite=False,
     return output
 
 
-def create_thumbnail(instance, thumbnail_remote_url, thumbnail_create_url=None, check_bbox=True, ogc_client=None):
-    if not ogc_client:
-        ogc_client = http_client
-    BBOX_DIFFERENCE_THRESHOLD = 1e-5
+def create_thumbnail(instance, thumbnail_remote_url, thumbnail_create_url=None,
+                     check_bbox=True, ogc_client=None, overwrite=False):
+    thumbnail_dir = os.path.join(settings.MEDIA_ROOT, 'thumbs')
+    thumbnail_name = 'layer-%s-thumb.png' % instance.uuid
+    thumbnail_path = os.path.join(thumbnail_dir, thumbnail_name)
 
-    if not thumbnail_create_url:
-        thumbnail_create_url = thumbnail_remote_url
+    if overwrite is True or storage.exists(thumbnail_path) is False:
+        if not ogc_client:
+            ogc_client = http_client
+        BBOX_DIFFERENCE_THRESHOLD = 1e-5
 
-    if check_bbox:
-        # Check if the bbox is invalid
-        valid_x = (
-            float(
-                instance.bbox_x0) -
-            float(
-                instance.bbox_x1)) ** 2 > BBOX_DIFFERENCE_THRESHOLD
-        valid_y = (
-            float(
-                instance.bbox_y1) -
-            float(
-                instance.bbox_y0)) ** 2 > BBOX_DIFFERENCE_THRESHOLD
-    else:
-        valid_x = True
-        valid_y = True
+        if not thumbnail_create_url:
+            thumbnail_create_url = thumbnail_remote_url
 
-    image = None
+        if check_bbox:
+            # Check if the bbox is invalid
+            valid_x = (
+                float(
+                    instance.bbox_x0) -
+                float(
+                    instance.bbox_x1)) ** 2 > BBOX_DIFFERENCE_THRESHOLD
+            valid_y = (
+                float(
+                    instance.bbox_y1) -
+                float(
+                    instance.bbox_y0)) ** 2 > BBOX_DIFFERENCE_THRESHOLD
+        else:
+            valid_x = True
+            valid_y = True
 
-    if valid_x and valid_y:
-        Link.objects.get_or_create(resource=instance.get_self_resource(),
-                                   url=thumbnail_remote_url,
-                                   defaults=dict(
-                                       extension='png',
-                                       name="Remote Thumbnail",
-                                       mime='image/png',
-                                       link_type='image',
+        image = None
+
+        if valid_x and valid_y:
+            Link.objects.get_or_create(resource=instance.get_self_resource(),
+                                       url=thumbnail_remote_url,
+                                       defaults=dict(
+                                           extension='png',
+                                           name="Remote Thumbnail",
+                                           mime='image/png',
+                                           link_type='image',
+                                           )
                                        )
-                                   )
-        Layer.objects.filter(id=instance.id).update(thumbnail_url=thumbnail_remote_url)
-        # Download thumbnail and save it locally.
-        resp, image = ogc_client.request(thumbnail_create_url)
-        if 'ServiceException' in image or resp.status < 200 or resp.status > 299:
-            msg = 'Unable to obtain thumbnail: %s' % image
-            logger.debug(msg)
-            # Replace error message with None.
-            image = None
+            Layer.objects.filter(id=instance.id) \
+                .update(thumbnail_url=thumbnail_remote_url)
+            # Download thumbnail and save it locally.
+            resp, image = ogc_client.request(thumbnail_create_url)
+            if 'ServiceException' in image or \
+               resp.status < 200 or resp.status > 299:
+                msg = 'Unable to obtain thumbnail: %s' % image
+                logger.debug(msg)
+                # Replace error message with None.
+                image = None
 
-    if image is not None:
-        filename = 'layer-%s-thumb.png' % instance.uuid
-        instance.save_thumbnail(filename, image=image)
+        if image is not None:
+            filename = 'layer-%s-thumb.png' % instance.uuid
+            instance.save_thumbnail(filename, image=image)
